# apps/docs/static/openapi/framework-live.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/billing-accounts": {
      "get": {
        "operationId": "BillingAccountController_getBillingAccounts",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Billing account status filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "required": false,
            "in": "query",
            "description": "Billing account number filter.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns billing accounts list."
          }
        },
        "summary": "List billing accounts",
        "tags": [
          "billing-accounts"
        ]
      }
    },
    "/api/billing-accounts/{id}": {
      "get": {
        "operationId": "BillingAccountController_getBillingAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Billing account identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns billing account details."
          }
        },
        "summary": "Get billing account by id",
        "tags": [
          "billing-accounts"
        ]
      }
    },
...
    "/api/checkout/{cartId}/addresses": {
      "post": {
        "operationId": "CheckoutController_setAddresses",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAddressesBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checkout addresses set."
          }
        },
        "summary": "Set checkout addresses",
        "tags": [
          "checkout"
        ]
      }
    },
...
    "/api/checkout/{cartId}/payment": {
      "post": {
        "operationId": "CheckoutController_setPayment",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPaymentBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checkout payment set."
          }
        },
        "summary": "Set checkout payment",
        "tags": [
          "checkout"
        ]
      }
    },
...
    "/api/checkout/{cartId}/summary": {
      "get": {
        "operationId": "CheckoutController_getCheckoutSummary",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "path",
            "description": "Optional locale code used to localize response values.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns checkout summary."
          }
        },
        "summary": "Get checkout summary",
        "tags": [
          "checkout"
        ]
      }
    },
    "/api/checkout/{cartId}/place-order": {
      "post": {
        "operationId": "CheckoutController_placeOrder",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order placement started."
          }
        },
        "summary": "Place order from cart",
        "tags": [
          "checkout"
        ]
      }
    },
    "/api/checkout/{cartId}/complete": {
      "post": {
        "operationId": "CheckoutController_completeCheckout",
        "parameters": [
          {
            "name": "cartId",
            "required": true,
            "in": "path",
            "description": "Cart identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteCheckoutBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checkout completed."
          }
        },
        "summary": "Complete checkout flow",
        "tags": [
          "checkout"
        ]
      }
    },
...
  },
  "info": {
    "title": "O2S API Harmonization HTTP API",
    "description": "Live API documentation generated from current NestJS controllers and DTOs.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Address": {
        "type": "object",
        "properties": {}
      },
...
      "SetAddressesBody": {
        "type": "object",
        "properties": {
          "shippingAddressId": {
            "type": "string",
            "description": "Saved shipping address identifier for authenticated users."
          },
          "shippingAddress": {
            "description": "Inline shipping address payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              }
            ]
          },
          "billingAddressId": {
            "type": "string",
            "description": "Saved billing address identifier for authenticated users."
          },
          "billingAddress": {
            "description": "Inline billing address payload.",
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              }
            ]
          },
          "notes": {
            "type": "string",
            "description": "Optional checkout notes."
          },
          "email": {
            "type": "string",
            "description": "Customer email, typically required for guest checkout."
          }
        }
      },
...
      "SetPaymentBody": {
        "type": "object",
        "properties": {
          "providerId": {
            "type": "string",
            "description": "Payment provider identifier."
          },
          "returnUrl": {
            "type": "string",
            "description": "URL used for redirect after successful payment flow."
          },
          "cancelUrl": {
            "type": "string",
            "description": "Optional URL used for redirect when payment flow is canceled."
          },
          "metadata": {
            "type": "object",
            "description": "Provider-specific metadata payload.",
            "additionalProperties": true
          }
        },
        "required": [
          "providerId",
          "returnUrl"
        ]
      },
      "PlaceOrderBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Customer email, required for guest checkout when not set earlier."
          }
        }
      },
      "CompleteCheckoutBody": {
        "type": "object",
        "properties": {
          "shippingAddressId": {
            "type": "string",
            "description": "Saved shipping address identifier for authenticated users."
          },
          "shippingAddress": {
            "description": "Inline shipping address payload (required for guest checkout).",
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              }
            ]
          },
          "billingAddressId": {
            "type": "string",
            "description": "Saved billing address identifier for authenticated users."
          },
          "billingAddress": {
            "description": "Inline billing address payload (required for guest checkout).",
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              }
            ]
          },
          "shippingMethodId": {
            "type": "string",
            "description": "Selected shipping method identifier."
          },
          "paymentProviderId": {
            "type": "string",
            "description": "Payment provider identifier."
          },
          "returnUrl": {
            "type": "string",
            "description": "URL used for redirect after successful payment flow."
          },
          "cancelUrl": {
            "type": "string",
            "description": "Optional URL used for redirect when payment flow is canceled."
          },
          "notes": {
            "type": "string",
            "description": "Optional checkout notes."
          },
          "email": {
            "type": "string",
            "description": "Customer email required for guest checkout."
          },
          "metadata": {
            "type": "object",
            "description": "Provider-specific metadata payload.",
            "additionalProperties": true
          }
        },
        "required": [
          "paymentProviderId",
          "returnUrl"
        ]
      },
...
    }
  }
}